+++ Milonchik/DetailViewController.swift
@@ -13,6 +13,16 @@ final class DetailViewController: NSViewController, StateResponding {
 
     @IBOutlet var detailView: DetailView!
 
+    lazy var htmlTemplate: String = {
+        guard
+            let htmlFile = Bundle.main.path(forResource: "DefinitionDetail", ofType: "html"),
+            let htmlTemplate = try? String(contentsOfFile: htmlFile, encoding: .utf8)
+        else {
+            preconditionFailure("Definition detail HTML template file is missing or corrupt")
+        }
+        return htmlTemplate
+    }()
+
     override func viewDidLoad() {
         //FIXME: white flash
         detailView.isHidden = true
@@ -32,49 +42,43 @@ final class DetailViewController: NSViewController, StateResponding {
     }
 
     private func loadHTML(with content: HTMLContent) {
-        guard
-            let htmlFile = Bundle.main.path(forResource: "DefinitionDetail", ofType: "html"),
-            let htmlTemplate = try? String(contentsOfFile: htmlFile, encoding: .utf8)
-        else {
-            preconditionFailure("Definition detail HTML template file is missing or corrupt")
-        }
-
         let htmlBody: String!
         switch content {
         case .placeholder(let message):
             htmlBody = "<p class=\"placeholder\">\(message)</p>"
         case .definition(let value):
             let partOfSpeech = value.partOfSpeech ?? ""
-            let translations = value.translations.map({"<li>\($0)</li>"}).joined()
-            let inflections = value.inflections.map({"<li>\($0.contents) | <span>\($0.kind)</span></li>"}).joined()
-            let samples = value.samples.map({ "<li>\($0)</li>" }).joined()
-            let synonyms = value.synonyms.map({ "<li>\($0)</li>" }).joined()
+
+            let translations = value.translations.map({"<li>\($0)</li>"})
+            let inflections = value.inflections.map({"<li>\($0.contents)<span> | \($0.kind)</span></li>"})
+            let samples = value.samples.map({ "<li>\($0)</li>" })
+            let synonyms = value.synonyms.joined(separator: ", ")
+
             let headers = makeHeaders(for: value.inputLanguage)
             let synonymsTitle = synonyms.isEmpty ? "" : headers.synonyms
             let inflectionsTitle = inflections.isEmpty ? "" : headers.inflections
             let samplesTitle = samples.isEmpty ? "" : headers.samples
+            // 1) word + part of speech, 2) translations, 3) inflections, 4) synonyms, 5) samples
             htmlBody = """
-            <!-- word + part of speech -->
             <h1 dir="auto">\(value.inputWord)<span> \(partOfSpeech)</span></h1>
-            <!-- translations -->
-            <ul dir="auto">\(translations)</ul>
-            <!-- inflections -->
-            <h2 dir="auto">\(inflectionsTitle)</h2>
-            <ul dir="auto">\(inflections)</ul>
-            <!-- synonyms -->
-            <h2 dir="auto">\(synonymsTitle)</h2>
+            <ul dir="auto" \(translations.count == 1 ? "class=\"singular\"" : "")>\(translations.joined())</ul>
+
+            <h2 dir="auto" \(inflections.isEmpty ? "class=\"hidden\"" : "")>\(inflectionsTitle)</h2>
+            <ul dir="auto" \(inflections.count == 1 ? "class=\"singular\"" : "")>\(inflections.joined())</ul>
+
+            <h2 dir="auto" \(synonyms.isEmpty ? "class=\"hidden\"" : "")>\(synonymsTitle)</h2>
             <p dir="auto">\(synonyms)</p>
-            <!-- samples -->
-            <h2 dir="auto">\(samplesTitle)</h2>
-            <ul dir="auto">\(samples)</ul>
+
+            <h2 dir="auto" \(samples.isEmpty ? "class=\"hidden\"" : "")>\(samplesTitle)</h2>
+            <ul dir="auto" \(samples.count == 1 ? "class=\"singular\"" : "")>\(samples.joined())</ul>
             """
         }
         detailView.loadHTMLString(htmlTemplate.replacingOccurrences(of: "$BODY", with: htmlBody), baseURL: nil)
 
         //FIXME: white flash
         if detailView.isHidden {
-            DispatchQueue.main.asyncAfter(deadline: .now() + 0.5) {
-                self.detailView.isHidden = false
+            DispatchQueue.main.asyncAfter(deadline: .now() + 0.7) {
+                self.detailView.animator().isHidden = false
             }
         }
     }