+++ MilonchikUITests/MilonchikUITests.swift
@@ -21,26 +21,34 @@ class MilonchikUITests: XCTestCase {
         sut.terminate()
     }
 
-    func testTypingQueryAndSelectingDefinition() throws {
-        let window = sut.children(matching: .window).firstMatch
-        let searchSearchField = window.toolbars.searchFields.firstMatch
-        searchSearchField.click()
-        searchSearchField.typeText("apple")
-        window/*@START_MENU_TOKEN@*/.tables.staticTexts["apple butter"]/*[[".splitGroups",".scrollViews.tables",".tableRows",".cells.staticTexts[\"apple butter\"]",".staticTexts[\"apple butter\"]",".tables"],[[[-1,5,2],[-1,1,2],[-1,0,1]],[[-1,5,2],[-1,1,2]],[[-1,4],[-1,3],[-1,2,3]],[[-1,4],[-1,3]]],[0,0]]@END_MENU_TOKEN@*/.click()
-        window.splitGroups.children(matching: .group).element.click()
-        XCTAssertTrue(window.webViews.staticTexts["ממרח תפוחים"].exists)
+    func testSearchFlow() throws {
+        let searchField = sut.searchFields.firstMatch
+        searchField.click()
+        searchField.typeText("apple")
+        sut.staticTexts["apple butter"].click()
+        XCTAssertTrue(sut.webViews.staticTexts["ממרח תפוחים"].exists)
         // window title reflects search results
-        XCTAssertTrue(window.title != "Milonchik")
+        XCTAssertTrue(sut.windows.firstMatch.title != "Milonchik")
         // navigation the table from the text field
-        sut.typeKey("f", modifierFlags: [.command, .alternate])
         sut.typeKey(.downArrow, modifierFlags: .function)
-        XCTAssertTrue(window.webViews.staticTexts["גון ירוק תפוח"].exists)
+        XCTAssertTrue(sut.staticTexts["גון ירוק תפוח"].exists)
     }
 
-    func testWindowTabbing() throws {
-        // tabbing
+    func testFocusSearchWithHotkey() throws {
+        sut.typeKey("f", modifierFlags: [.command, .alternate])
+        XCTAssertTrue(sut.searchFields.firstMatch.isSelected)
+    }
+
+    func testTabbing() throws {
         sut.typeKey("t", modifierFlags: .command)
-        XCTAssertTrue(sut.tabs.count == 2)
+        XCTAssertTrue(sut.tabs.count > 1)
+    }
+
+    func testNoReloadContextMenu() throws {
+        let webView = sut.webViews
+        let placeholder = webView.staticTexts.firstMatch
+        placeholder.coordinate(withNormalizedOffset: CGVector(dx: 5, dy: 5)).click()
+        XCTAssertTrue(webView.menuItems.count == 0)
     }
 
     // func testLaunchPerformance() throws {